Vendorize instead of godep.
