+++ .gitmodules
@@ -1,0 +1,18 @@
+[submodule "vendor/src/github.com/davecgh/go-spew"]
+	path = vendor/src/github.com/davecgh/go-spew
+	url = https://github.com/davecgh/go-spew
+[submodule "vendor/src/github.com/fsouza/go-dockerclient"]
+	path = vendor/src/github.com/fsouza/go-dockerclient
+	url = https://github.com/fsouza/go-dockerclient
+[submodule "vendor/src/github.com/n1koo/gocron"]
+	path = vendor/src/github.com/n1koo/gocron
+	url = https://github.com/n1koo/gocron
+[submodule "vendor/src/github.com/pmezard/go-difflib"]
+	path = vendor/src/github.com/pmezard/go-difflib
+	url = https://github.com/pmezard/go-difflib
+[submodule "vendor/src/github.com/stretchr/testify"]
+	path = vendor/src/github.com/stretchr/testify
+	url = https://github.com/stretchr/testify
+[submodule "vendor/src/github.com/Sirupsen/logrus"]
+	path = vendor/src/github.com/Sirupsen/logrus
+	url = https://github.com/Sirupsen/logrus
