+++ script/setup
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -xe
+
+go get github.com/constabulary/gb/...
+git submodule update --init --recursive
