+++ script/test
@@ -1,3 +1,4 @@
 #!/bin/bash
+set -xe
 
-go test -v -race ./...
+gb test ./...
